populate_backlog: re-auto-categorize as appropriate
If the auto categorization would place it in its current location in the
backlog given its old path, then move it to a new location based on auto
categorization, rather than leaving it where it is.

As one example, if the game lives in the area where my broken games are, and
as such is categorized as broken, and I fix the game and move it to my usual
games area, I want it re-categorized out of the broken area.

Signed-off-by: Christopher Larson <kergoth@gmail.com>